requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==3.6.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==0.15.0
pytz==2024.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
